Move tests to Auth namespace
